+++ ParseSwift.playground/Pages/2 - Finding Objects.xcplaygroundpage/Contents.swift
@@ -21,7 +21,7 @@ struct GameScore: ParseObject {
 
     //: Your own properties.
     var score: Int?
-    var timeStamp = Date()
+    var timeStamp: Date? = Date()
     var oldScore: Int?
 }